fix(evals): decode plugin manifests that use quoted sampling/duration strings (#161)
Phase-A templates authored `sampling: "0.1"` and `interval: 60s` as
quoted strings so Helm template interpolation would not double-quote
the values. The Go struct declared those fields as `float64` and
`time.Duration`, and `go.yaml.in/yaml/v3` is strict about scalar
typing — so installing the Langfuse template from the console rendered
"decode plugin manifest: yaml: unmarshal errors: line 13: cannot
unmarshal !!str `0.1` into float64" on a previously-OK manifest once
the user toggled a single field.

Add two YAML-aware wrappers around the scalar fields:

- 'evalplugin.SamplingFraction' (float64 alias with UnmarshalYAML)
  accepts unquoted float, quoted float, and percent strings (`25%`).
  Bare integers are NOT silently coerced — they fall through to the
  same [0,1] validator rejection that already protected against
  unintended oversampling, so "sampling: 15" still fails loudly.

- 'evalplugin.Duration' (time.Duration alias with UnmarshalYAML)
  accepts Go strings (`60s`), ms strings (`1500ms`), and bare numeric
  seconds (so Helm `.Files.Get` interpolation that always yields a
  string still resolves). Out-of-range or unparseable values return
  explicit errors rather than the cryptic `cannot unmarshal` form.

Untangle the type changes through:

  internal/evalplugin/types.go              (SamplingFraction, Duration)
  internal/evalplugin/validate.go           (unwrap via float64()/Std())
  internal/evaluators/external/multi.go     (pass float64 into sampler)
  internal/evaluators/external/collector.go (use Std() for interval)
  internal/evalplugin/plugin_test.go        (multi-form cases + quoted
                                             sampling regression)
  web/src/pages/EvalPlugins.tsx             (default manifest uses
                                             unquoted `sampling: 0.1`
                                             for clarity)

Also exclude templates/eval-plugins/ from the helm chart layer's
template scan via .helmignore. The two reference YAMLs in that
directory are not referenced from any active template; the helm chart
parser still reads every .yaml under templates/ as a candidate
template and fails on the literal `\n` inside the langfuse manifest
during dry-runs. Excluding the directory keeps the chart renderable
while the example files remain visible for human reference.

+++ internal/evaluators/external/multi.go
@@ -41,7 +41,7 @@ func (m *MultiEvaluator) Name() string { return "external" }
 func (m *MultiEvaluator) Evaluate(ctx context.Context, t observability.Trace) ([]evals.Score, error) {
 	enabled := m.reg.Enabled()
 	for _, rec := range enabled {
-		if !sampleTrace(rec.Plugin.Spec.Send.Sampling) {
+		if !sampleTrace(float64(rec.Plugin.Spec.Send.Sampling)) {
 			continue
 		}
 		_ = m.dispatcher.Dispatch(ctx, t, rec.Plugin)
